+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
    - Added keyword ignore_empty_files to pysat.Instrument and Files objects
      to filter out empty files from the stored file list
    - Added slice and list ability to meta
+   - Converted all print statements to logging statements
    - Updated cleaning routines for C/NOFS IVM
    - Added S4 scintillation data to the cosmic-gps instrument
    - pysat no longer creates a default data directory. User must specify location.